requirements.txt
transformers
torch
pandas
python-dotenv
